requirements.txt
pycryptodomex
telethon
flask
aiofiles
aiohttp
